building moment

+++ Source/BKMusicCore/Public/M2SoundGraphData.h
@@ -188,7 +188,7 @@ struct FLinkedMidiEvents
 	double StartTime = 0.0;
 
 	UPROPERTY(VisibleAnywhere, Category = "unDAW|Midi Data", BlueprintReadOnly)
-	float NoteVelocity = 0.0f;
+	uint8 NoteVelocity = 0.0f;
 
 	UPROPERTY(VisibleAnywhere, Category = "unDAW|Midi Data")
 	float cornerRadius = 0.0f;
